debt: playground: extract logs.js (live events + clipboard export)
Live System Logs + Copy System Logs combined into one module — they
share the transitionLog array, pagination cursor, and mode-change
detector. transitionLog is exported as a live binding so main.js's
simLoop, resetSim, restoreBootstrapSnapshot and mode-switch code
can still mutate the same array directly (push/unshift/pop/length).

Exposes resetEventsState() as the setter for the pagination vars
that were previously assigned inline in main.js's clearLiveDisplay.

main.js exports added: model, params, MODE_INFO, formatTimeOfDay —
consumed by logs.js for clipboard-text composition.

main.js: 2328 → 1992 lines. npm test:unit 782/782 pass. knip clean.
